# platform_api/openapi.yaml
openapi: 3.0.3
info:
  title: Mwalimu Platform API
  version: 1.0.0
  description: Identity, institution, membership, and library API for the Mwalimu
    platform.
paths:
...
  /api/v1/libraries/{library_pk}/resources/:
    get:
      operationId: api_v1_libraries_resources_list
      description: |-
        View set for resource management within a library.

        Library administrators and institution administrators may create, update,
        and delete resources. Authorized library users may list and retrieve
        resources. The original binary is stored in S3-compatible object storage;
        only metadata is kept in PostgreSQL.
      parameters:
      - in: path
        name: library_pk
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - api
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResourceList'
          description: ''
    post:
      operationId: api_v1_libraries_resources_create
      description: |-
        Upload a resource into the library.

        Only library managers may upload resources.
      parameters:
      - in: path
        name: library_pk
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ResourceRequest'
        required: true
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
          description: ''
  /api/v1/libraries/{library_pk}/resources/{id}/:
    get:
      operationId: api_v1_libraries_resources_retrieve
      description: |-
        View set for resource management within a library.

        Library administrators and institution administrators may create, update,
        and delete resources. Authorized library users may list and retrieve
        resources. The original binary is stored in S3-compatible object storage;
        only metadata is kept in PostgreSQL.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: library_pk
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
          description: ''
    patch:
      operationId: api_v1_libraries_resources_partial_update
      description: Only library managers may partially update resource metadata.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: library_pk
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedResourceRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedResourceRequest'
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
          description: ''
    delete:
      operationId: api_v1_libraries_resources_destroy
      description: |-
        Delete the resource and its stored object.

        Storage is deleted first so the system of record never points to a
        missing object. If storage deletion fails, the database record is kept.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: library_pk
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - api
      security:
      - jwtAuth: []
      - BearerAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    PaginatedResourceList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Resource'
...
    PatchedResourceRequest:
      type: object
      description: Serializer for resource metadata.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        resource_type:
          $ref: '#/components/schemas/ResourceTypeEnum'
    Resource:
      type: object
      description: Serializer for resource metadata.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        library:
          allOf:
          - $ref: '#/components/schemas/ResourceLibrary'
          readOnly: true
        name:
          type: string
          maxLength: 255
        resource_type:
          $ref: '#/components/schemas/ResourceTypeEnum'
        original_filename:
          type: string
          readOnly: true
        content_type:
          type: string
          readOnly: true
        size:
          type: integer
          readOnly: true
        object_key:
          type: string
          readOnly: true
        checksum:
          type: string
          readOnly: true
          description: SHA-256 hex digest of the original content.
        status:
          allOf:
          - $ref: '#/components/schemas/ResourceStatusEnum'
          readOnly: true
        created_by:
          allOf:
          - $ref: '#/components/schemas/ResourceUser'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - checksum
      - content_type
      - created_at
      - created_by
      - id
      - library
      - name
      - object_key
      - original_filename
      - resource_type
      - size
      - status
      - updated_at
    ResourceLibrary:
      type: object
      description: Minimal library representation nested inside resource responses.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 255
        slug:
          type: string
          maxLength: 255
          pattern: ^[-a-zA-Z0-9_]+$
      required:
      - id
      - name
      - slug
...
    ResourceRequest:
      type: object
      description: Serializer for resource metadata.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        resource_type:
          $ref: '#/components/schemas/ResourceTypeEnum'
      required:
      - name
      - resource_type
    ResourceStatusEnum:
      enum:
      - pending
      - uploading
      - ready
      - failed
      - archived
      type: string
      description: |-
        * `pending` - Pending
        * `uploading` - Uploading
        * `ready` - Ready
        * `failed` - Failed
        * `archived` - Archived
    ResourceTypeEnum:
      enum:
      - pdf
      - docx
      - txt
      type: string
      description: |-
        * `pdf` - PDF
        * `docx` - DOCX
        * `txt` - Text
    ResourceUser:
      type: object
      description: Minimal user representation nested inside resource responses.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          format: email
          maxLength: 254
      required:
      - email
      - id
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT